feat: big-map: refactor big-map-view service
